Added fix for timing issue with mock HTTP server used in tests
Previously, various tests could fail due to the mock HTTP server not
being ready to handle requests before a test caused a request to be
made.

In this commit, a 'healthcheck' endpoint is added to the mock
server. Requests are made to the 'healthcheck' endpoint repeatedly after
the initial server startup until a 200 OK is returned, signifying that
the server is ready to receive the additional requests that a test may
make.

Some additional test cleanup was also added for closing resources
appropriately in order to avoid the display of error messages shown when
running in Python 3.7.